Change int -> uint8_t in tsa.h, tsa.c, dsa.c

+++ MVbR2TGavr/atmega32a/atmega32a/dsa.c
@@ -17,7 +17,7 @@ Last Modified: 2022/01/25
 void printList (struct Node* headnode) {
 
     struct Node* p = headnode;
-    int count = 0;
+    uint8_t count = 0;
     while ( p->_childpresent != 0 ) {
         printf ("\nNode %d x coordinate: %d", count, p->pixel._coord._x);
         printf ("\nNode %d y coordinate: %d", count, p->pixel._coord._y);
@@ -28,7 +28,7 @@ void printList (struct Node* headnode) {
 
 // append new node at the end of the list
 void appendNode ( struct Node* currentnode, struct Pixel pixeldata) {
-    
+
     //create a node in the heap
     struct Node* _node = ( struct Node* ) malloc (sizeof (struct Node));
     _node->_childpresent = 0;
@@ -43,10 +43,10 @@ void appendNode ( struct Node* currentnode, struct Pixel pixeldata) {
 
 // latest implementation of the append fucntion
 void appendNodeToLinkedList (struct Node** headnode, struct Node** currentnode, struct Pixel pixeldata) {
-    
+
     // create a node in the heap
     struct Node* _node = ( struct Node* ) malloc (sizeof (struct Node));
-    
+
     // newly created node is the last node of the linked list. Make its nextnode NULL
     _node->nextnode = NULL;
 
@@ -56,27 +56,27 @@ void appendNodeToLinkedList (struct Node** headnode, struct Node** currentnode, 
     // check whether the linked list is empty at the moment of fucntion call
     if ( (*headnode) == NULL ) {
         // if list is empty make newly created node the head.
-        (*headnode) = _node; 
+        (*headnode) = _node;
     } else 	{
         // if the list is not empty make connection between current node and newly created node
         (*currentnode)->nextnode = _node;
     }
-    
-    // point the currnt node pointer to the newly ce\reated node   
+
+    // point the currnt node pointer to the newly ce\reated node
     (*currentnode) = _node;
 }
 
 void printLinkedList (struct Node* headnode) {
-    
+
     // if the headnode is null that means there is no node in the list.
     if ( headnode == NULL ) {
         printf ("\nLinked List is empty!");
         return;
     }
-    
+
     // creating a local variable to keep the pointer
     struct Node* _tempnode = headnode;
-    
+
     // printing the linked list's elements if it is not empty
     printf ("\n{");
     while ( _tempnode != NULL ) {
